Lock in dictresize, not in callers

+++ Include/cpython/pyatomic_msc.h
@@ -945,9 +945,9 @@ _Py_atomic_store_ssize_release(Py_ssize_t *obj, Py_ssize_t value)
 #if defined(_M_X64) || defined(_M_IX86)
     *(Py_ssize_t volatile *)obj = value;
 #elif defined(_M_ARM64)
-    __stlr64((unsigned __int64 volatile *)obj, (uintptr_t)value);
+    __stlr64((unsigned __int64 volatile *)obj, (unsigned __int64)value);
 #else
-#  error "no implementation of _Py_atomic_store_int_release"
+#  error "no implementation of _Py_atomic_store_ssize_release"
 #endif
 }
